Crafting finalized, equip update, other misc fixes and additions
